+++ zcash_client_backend/src/data_api/wallet.rs
@@ -2283,7 +2283,7 @@ where
                         Ok(SentTransactionOutput::from_parts(
                             output_index,
                             recipient,
-                            output.value,
+                            output.value(),
                             None,
                         ))
                     })